--- .scrutinizer.yml
@@ -1,161 +1,0 @@
-# language: php
-# .scrutinizer.yml
-before_commands:
-    - "composer install --prefer-dist --no-interaction"
-
-filter:
-    paths:
-        - src/*
-    excluded_paths:
-        - bin/*
-        - tests/*
-        - docs/*
-        - vendor/*
-        - _*
-
-tools:
-    external_code_coverage:
-        timeout: 600
-    php_code_coverage: false
-    php_code_sniffer:
-        enabled: true
-        config:
-            standard: 'PSR2'
-        filter:
-            paths:
-                - src/*
-            excluded_paths:
-                - bin/*
-                - tests/*
-                - docs/*
-                - vendor/*
-                - _*
-    php_cpd:
-        enabled: true
-        excluded_dirs:
-            - bin/*
-            - tests/*
-            - docs/*
-            - vendor/*
-            - _*
-    php_cs_fixer:
-        enabled: true
-        config:
-            level: all
-        filter:
-            paths:
-                - src/*
-    php_loc:
-        enabled: true
-        excluded_dirs:
-            - bin/*
-            - tests/*
-            - docs/*
-            - vendor/*
-            - _*
-    php_mess_detector:
-        enabled: true
-        filter:
-            paths:
-                - src/*
-    php_pdepend:
-        enabled: true
-        excluded_dirs:
-            - bin/*
-            - tests/*
-            - docs/*
-            - vendor/*
-            - _*
-    php_analyzer: true
-    php_analyzer:
-        filter:
-            paths:
-                - src/*
-    sensiolabs_security_checker: true
-
-checks:
-    php:
-        code_rating: true
-        variable_existence: true
-        verify_property_names: true
-        unused_variables: true
-        unreachable_code: true
-        sql_injection_vulnerabilities: true
-        single_namespace_per_use: true
-        security_vulnerabilities: true
-        require_php_tag_first: true
-        psr2_switch_declaration: true
-        psr2_class_declaration: false
-        remove_extra_empty_lines: true
-        remove_php_closing_tag: true
-        remove_trailing_whitespace: true
-        fix_use_statements:
-            remove_unused: true
-            preserve_multiple: false
-            preserve_blanklines: true
-            order_alphabetically: true
-        fix_php_opening_tag: true
-        fix_linefeed: true
-        fix_line_ending: true
-        fix_identation_4spaces: true
-        fix_doc_comments: true
-        use_self_instead_of_fqcn: true
-        uppercase_constants: true
-        simplify_boolean_return: true
-        return_doc_comments: true
-        properties_in_camelcaps: true
-        prefer_while_loop_over_for_loop: true
-        phpunit_assertions: true
-        parameters_in_camelcaps: true
-        optional_parameters_at_the_end: true
-        no_short_variable_names:
-            minimum: '3'
-        no_short_method_names:
-            minimum: '3'
-        no_new_line_at_end_of_file: true
-        no_long_variable_names:
-            maximum: '60'
-        no_goto: true
-        newline_at_end_of_file: true
-        line_length:
-            max_length: '120'
-        function_in_camel_caps: true
-        encourage_single_quotes: true
-        encourage_postdec_operator: true
-        avoid_perl_style_comments: true
-        avoid_multiple_statements_on_same_line: true
-        avoid_fixme_comments: true
-        align_assignments: true
-        return_doc_comment_if_not_inferrable: true
-        parameter_doc_comments: true
-        param_doc_comment_if_not_inferrable: true
-        more_specific_types_in_doc_comments: true
-        avoid_unnecessary_concatenation: true
-
-coding_style:
-    php:
-        spaces:
-            other:
-                after_type_cast: true
-        braces:
-            classes_functions:
-                class: new-line
-                function: new-line
-                closure: end-of-line
-            if:
-                opening: end-of-line
-            for:
-                opening: end-of-line
-            while:
-                opening: end-of-line
-            do_while:
-                opening: end-of-line
-            switch:
-                opening: end-of-line
-            try:
-                opening: end-of-line
-        upper_lower_casing:
-            keywords:
-                general: lower
-            constants:
-                true_false_null: lower
