+++ phpunit.xml.dist
@@ -1,0 +1,28 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<phpunit xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:noNamespaceSchemaLocation="http://schema.phpunit.de/4.5/phpunit.xsd"
+         backupGlobals="false"
+>
+
+
+    <testsuites>
+        <testsuite name="clickalicious\rng Test Suite">
+            <directory suffix="Test.php">./tests</directory>
+        </testsuite>
+    </testsuites>
+
+    <filter>
+        <whitelist>
+            <directory>./src</directory>
+            <exclude>
+                <directory>./bin</directory>
+                <directory>./docs</directory>
+                <directory>./vendor</directory>
+                <directory>./tests</directory>
+                <directory>./src/Rng/Autoloader.php</directory>
+                <directory>./src/Rng/Exception.php</directory>
+            </exclude>
+        </whitelist>
+    </filter>
+</phpunit>
