+++ update-gh-pages.sh
@@ -3,7 +3,7 @@ if [ $TRAVIS_PULL_REQUEST == 'false' ]; then
   echo -e "Starting to update gh-pages\n"
 
   #copy data we're interested in to other place
-  cp -R docs/coverage $HOME/coverage
+  cp -R build/html/coverage $HOME/coverage
 
   #go to home and setup git
   cd $HOME
